Starting iteration 3

+++ lib/item.rb
@@ -4,7 +4,7 @@ class Item
   attr_reader :name, :price
   def initialize(attributes)
     @name = attributes[:name]
-    @price = attributes[:price].delete('$').to_i
+    @price = attributes[:price].delete('$').to_f
   end
 
 end
